Fix install on new systems
On a new system where the package does not exist yet, the __version__
cannot be imported from mensa_ukon.version during setup as the package
is not yet there to be imported.
One solution would be to use a standalone VERSION file that contains
only the version number. However this makes importing the file
more difficult.
As I would rather have a more involved way to get to the version
during setup than during program execution, we execute the file
containing the version and store the variable into a separate name-
space.
This way, we get directly importable version information at runtime
and the same version at setup time.